--- src/workspacer.Shared/Keybind/Modes.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace workspacer
-
-{
-   public enum ModeDefaultBindings
-    {
-        None,
-        Minimal,
-        All
-    }
-}
