+++ src/workspacer/ConfigContext.cs
@@ -29,7 +29,7 @@ namespace workspacer
         public IWorkspaceContainer WorkspaceContainer { get; set; }
         public IWindowRouter WindowRouter { get; set; }
         public IMonitorContainer MonitorContainer { get; set; }
-        public KeyMode Mode { get; set; }
+       
         private System.Timers.Timer _timer;
         private PipeServer _pipeServer;
         private Func<ILayoutEngine[]> _defaultLayouts;
@@ -57,13 +57,11 @@ namespace workspacer
             Workspaces = new WorkspaceManager(this);
             Windows = new WindowsManager();
             Keybinds = new KeybindManager(this);
-            
+
             WorkspaceContainer = new WorkspaceContainer(this);
             WindowRouter = new WindowRouter(this);
             MonitorContainer = new NativeMonitorContainer();
 
-            var mode = new KeyMode(this, "test");
-            Keybinds.SetMode(mode);
 
             Windows.WindowCreated += Workspaces.AddWindow;
             Windows.WindowDestroyed += Workspaces.RemoveWindow;
@@ -72,7 +70,7 @@ namespace workspacer
             // ignore watcher windows in workspacer
             WindowRouter.AddFilter((window) => !(window.ProcessId == _pipeServer.WatcherProcess.Id));
         }
-       
+      
         public void ConnectToWatcher()
         {
             _pipeServer.Start();
@@ -89,6 +87,7 @@ namespace workspacer
                 Logger.ConsoleLogLevel = value;
             }
         }
+
 
         public LogLevel FileLogLevel
         {
